+++ pyobs/images/processors/offsets/__init__.py
@@ -4,5 +4,6 @@ Offsets
 """
 
 from .offsets import Offsets
+from .astrometry import AstrometryOffsets
 from .nstar import NStarOffsets
 from .projected import ProjectedOffsets