+++ pyobs/utils/simulation/camera.py
@@ -9,10 +9,14 @@ from astropy.wcs import WCS
 from astropy.io import fits
 from photutils.datasets import make_gaussian_prf_sources_image
 from photutils.datasets import make_noise_image
+import logging
 
 from pyobs.object import Object
 from pyobs.utils.enums import ImageFormat
 from pyobs.images import Image
+
+
+log = logging.getLogger(__name__)
 
 
 class SimCamera(Object):
@@ -20,7 +24,7 @@ class SimCamera(Object):
     __module__ = 'pyobs.utils.simulation'
 
     def __init__(self, world: 'SimWorld', image_size: Tuple[int, int] = None, pixel_size: float = 0.015,
-                 images: str = None, *args, **kwargs):
+                 images: str = None, max_mag: float = 20., *args, **kwargs):
         """Inits a new camera.
 
         Args:
@@ -28,6 +32,7 @@ class SimCamera(Object):
             image_size: Size of image.
             pixel_size: Square pixel size in mm.
             images: Filename pattern (e.g. /path/to/*.fits) for files to return instead of simulated images.
+            max_mag: Maximum magnitude for sim.
         """
         Object.__init__(self, *args, **kwargs)
 
@@ -41,6 +46,7 @@ class SimCamera(Object):
         self.image_format = ImageFormat.INT16
         self.images = [] if images is None else sorted(glob.glob(images)) \
             if '*' in images or '?' in images else [images]
+        self._max_mag = max_mag
 
         # private stuff
         self._catalog = None
@@ -151,24 +157,61 @@ class SimCamera(Object):
         # finished
         return hdr
 
-    def _get_catalog(self):
+    def _get_catalog(self, fov):
         """Returns GAIA catalog for current telescope coordinates."""
-        from astroquery.gaia import Gaia
-
         # get catalog
         if self._catalog_coords is None or self._catalog_coords.separation(self.telescope.real_pos) > 10. * u.arcmin:
-            self._catalog = Gaia.query_object_async(coordinate=self.telescope.real_pos, radius=1. * u.deg)
+            from astroquery.utils.tap import TapPlus
+
+            # get coordinates
+            coords = self.telescope.real_pos
+
+            # query TAP
+            tap = TapPlus(url="https://gea.esac.esa.int/tap-server/tap")
+            query = self._get_gaia_query(coords.ra.degree, coords.dec.degree, fov * 1.5)
+            job = tap.launch_job(query)
+
+            # get result table
+            self._catalog = job.get_results()
+
         return self._catalog
+
+    def _get_gaia_query(self, ra, dec, radius):
+        # define query
+        return f"""
+                SELECT
+                  TOP 1000
+                  DISTANCE(
+                    POINT('ICRS', ra, dec),
+                    POINT('ICRS', {ra}, {dec})
+                  ) as dist,
+                  ra, dec, phot_g_mean_flux, phot_g_mean_mag
+                FROM
+                  gaiadr2.gaia_source
+                WHERE
+                  1 = CONTAINS(
+                    POINT('ICRS', ra, dec),
+                    CIRCLE('ICRS', {ra}, {dec}, {radius})
+                  )
+                  AND phot_g_mean_mag < {self._max_mag}
+                ORDER BY
+                  phot_g_mean_mag ASC
+                """
 
     def _get_sources_table(self, exp_time: float):
         """Create sources table."""
 
-        # get catalog
-        cat = self._get_catalog()
-
         # calculate cdelt1/2
         tmp = 360. / (2. * np.pi) * self.pixel_size / self.telescope.focal_length
         cdelt1, cdelt2 = tmp * self.binning[0], tmp * self.binning[1]
+        log.info('Plate scale is %.2f"/px, image size is %.2f\'x%.2f\'.',
+                 cdelt1*3600, cdelt1*60*self.window[2], cdelt2*60*self.window[3])
+
+        # FoV
+        fov = np.max(cdelt2 * np.array(self.full_frame[2:]))
+
+        # get catalog
+        cat = self._get_catalog(fov)
 
         # create WCS
         w = WCS(naxis=2)
@@ -177,17 +220,17 @@ class SimCamera(Object):
         w.wcs.crval = [self.telescope.real_pos.ra.degree, self.telescope.real_pos.dec.degree]
         w.wcs.ctype = ["RA---TAN", "DEC--TAN"]
 
-        # set fwhm to 2" in pixels
-        fwhm = 2. / 3600. / cdelt1
+        # set sigma to 4" FWHM in pixels
+        fwhm = 10. / 3600. / cdelt1 / 2.3548
 
         # convert world to pixel coordinates
         cat['x'], cat['y'] = w.wcs_world2pix(cat['ra'], cat['dec'], 0)
 
         # get columns
-        sources = cat['x', 'y', 'phot_g_mean_flux']
-        sources.rename_columns(['x', 'y', 'phot_g_mean_flux'], ['x_0', 'y_0', 'amplitude'])
+        sources = cat['x', 'y', 'phot_g_mean_flux', 'phot_g_mean_mag']
+        sources.rename_columns(['x', 'y', 'phot_g_mean_flux'], ['x_0', 'y_0', 'flux'])
         sources.add_column([fwhm] * len(sources), name='sigma')
-        sources['amplitude'] *= exp_time
+        sources['flux'] *= exp_time
 
         # finished
         return sources