+++ pyobs/utils/skyflats/flatfielder.py
@@ -9,7 +9,7 @@ from astropy.time import TimeDelta
 import numpy as np
 from py_expression_eval import Parser
 
-from pyobs.interfaces import ITelescope, ICamera, IFilters, ICameraBinning, ICameraWindow, ICameraExposureTime, \
+from pyobs.interfaces import ITelescope, ICamera, IFilters, IBinning, IWindow, IExposureTime, \
     IImageType
 from pyobs.object import get_object
 from pyobs.utils.enums import ImageType
@@ -135,7 +135,7 @@ class FlatFielder:
         self._cur_filter: Optional[str] = None
         self._cur_binning: Optional[int] = None
 
-    def __call__(self, telescope: ITelescope, camera: Union[ICamera, ICameraExposureTime], filters: IFilters,
+    def __call__(self, telescope: ITelescope, camera: Union[ICamera, IExposureTime], filters: IFilters,
                  filter_name: str, count: int = 20, binning: int = 1) -> State:
         """Calls next step in state machine.
 
@@ -152,7 +152,7 @@ class FlatFielder:
         """
 
         # camera must support exposure times
-        if not isinstance(camera, ICameraExposureTime):
+        if not isinstance(camera, IExposureTime):
             raise ValueError('Camera must support exposure times.')
 
         # store
@@ -214,7 +214,7 @@ class FlatFielder:
             log.info('Flat-field time is still coming, keep going...')
             return True
 
-    def _init_system(self, telescope: ITelescope, camera: Union[ICamera, ICameraExposureTime], filters: IFilters):
+    def _init_system(self, telescope: ITelescope, camera: Union[ICamera, IExposureTime], filters: IFilters):
         """Initialize whole system."""
 
         # which twilight are we in?
@@ -229,7 +229,7 @@ class FlatFielder:
             return
 
         # set binning
-        if isinstance(camera, ICameraBinning):
+        if isinstance(camera, IBinning):
             log.info('Setting binning to %dx%d...', self._cur_binning, self._cur_binning)
             camera.set_binning(self._cur_binning, self._cur_binning)
 
@@ -251,7 +251,7 @@ class FlatFielder:
         log.info('Waiting for flat-field time...')
         self._state = FlatFielder.State.WAITING
 
-    def _get_bias(self, camera: Union[ICamera, ICameraExposureTime]) -> float:
+    def _get_bias(self, camera: Union[ICamera, IExposureTime]) -> float:
         """Take bias image to determine bias level.
 
         Returns:
@@ -260,7 +260,7 @@ class FlatFielder:
         log.info('Taking BIAS image to determine median level...')
 
         # set full frame
-        if isinstance(camera, ICameraWindow):
+        if isinstance(camera, IWindow):
             full_frame = camera.get_full_frame().wait()
             camera.set_window(*full_frame).wait()
 
@@ -355,7 +355,7 @@ class FlatFielder:
             # otherwise it seems that we're in the middle of flat-fielding time
             return 0
 
-    def _testing(self, camera: Union[ICamera, ICameraExposureTime]):
+    def _testing(self, camera: Union[ICamera, IExposureTime]):
         """Take flat-fields but don't store them."""
 
         # set window
@@ -383,13 +383,13 @@ class FlatFielder:
             log.info('Missed flat-fielding time, finish task...')
             self._state = FlatFielder.State.FINISHED
 
-    def _set_window(self, camera: Union[ICamera, ICameraExposureTime], testing: bool):
+    def _set_window(self, camera: Union[ICamera, IExposureTime], testing: bool):
         """Set camera window.
 
         Args:
             testing: Whether we're in testing mode or not.
         """
-        if isinstance(camera, ICameraWindow):
+        if isinstance(camera, IWindow):
             # get full frame
             left, top, width, height = camera.get_full_frame().wait()